// backend/NodeJs/src/tricks/entities/trick.interface.ts
export interface TricksI {
  id?: number;
  name: string | null;
  point: number | null;
  velocity: number | null;
  dateAdd: Date | null;
  authorId: number | null;
  mapId: number | null;
  // completes: Completes[];
  // routes: Routes[];
  // swrs: Swr[];
  // author: Players;
  // twrs: Twr[];
  // map: Maps;
}